refactor: move model classes back to `tables` package
