+++ services/rsu-info-bridge/src/main/java/com/trihydro/rsuinfobridge/models/tables/RsuIntersection.java
@@ -1,4 +1,4 @@
-package com.trihydro.rsuinfobridge.testutil.model;
+package com.trihydro.rsuinfobridge.models.tables;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
