+++ services/rsu-info-bridge/src/test/java/com/trihydro/rsuinfobridge/testutil/repository/RsuIntersectionRepository.java
@@ -3,7 +3,7 @@ package com.trihydro.rsuinfobridge.testutil.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.trihydro.rsuinfobridge.testutil.model.RsuIntersection;
+import com.trihydro.rsuinfobridge.models.tables.RsuIntersection;
 
 @Repository
 public interface RsuIntersectionRepository extends JpaRepository<RsuIntersection, Integer> {
