+++ services/rsu-info-bridge/src/test/java/com/trihydro/rsuinfobridge/testutil/repository/RsuOrganizationRepository.java
@@ -3,7 +3,7 @@ package com.trihydro.rsuinfobridge.testutil.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.trihydro.rsuinfobridge.testutil.model.RsuOrganization;
+import com.trihydro.rsuinfobridge.models.tables.RsuOrganization;
 
 @Repository
 public interface RsuOrganizationRepository extends JpaRepository<RsuOrganization, Integer> {
